backend/requirements.txt
aiofiles==23.1.0
anyio==3.6.2
asyncpg==0.27.0
attrs==22.2.0
bcrypt==4.0.1
certifi==2022.12.7
cffi==1.15.1
click==8.1.3
colorama==0.4.6
cryptography==3.4.8
dnspython==2.3.0
email-validator==1.3.1
exceptiongroup==1.1.1
fastapi==0.93.0
fastapi-jwt-auth==0.5.0
greenlet==2.0.2
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.4
iniconfig==2.0.0
Jinja2==3.1.2
MarkupSafe==2.1.2
packaging==23.0
passlib==1.7.4
pluggy==1.0.0
psycopg2==2.9.5
pycparser==2.21
pydantic==1.10.6
PyJWT==1.7.1
pytest==7.2.2
pytest-anyio==0.0.0
python-dotenv==1.0.0
rfc3986==1.5.0
sniffio==1.3.0
SQLAlchemy==1.4.41
sqlalchemy2-stubs==0.0.2a32
sqlmodel==0.0.8
starlette==0.25.0
tomli==2.0.1
typing_extensions==4.5.0
uvicorn==0.20.0
